implemented speed up output dir select
