+++ saves/save.json
@@ -1,4 +1,6 @@
 {
     "recent_files": [
+    ],
+    "default_output_location": [
     ]
 }
